+++ src/types/IFilter.ts
@@ -2,5 +2,6 @@ import * as PIXI from "pixi.js";
 
 export interface IFilter {
     container: PIXI.Container;
-    flashback: boolean;
+    flashback?: boolean;
+    sick?: { container?: PIXI.Container | null; show?: boolean };
 }