+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 5.5.0
 Release date: July 8, 2025
-- Added flashback filter
+- Added filters (Flashback and Nausea)
 - Added background bookmark function
 - Allow 2 to 6 easy switch inputs
 - Allow hiding of models and dialogue temporarily for scene text